move orientation to options object closes #281
