+++ projects/ngx-d3timeline/src/lib/options.ts
@@ -1,0 +1,5 @@
+import { Orientation } from './orientation';
+
+export interface Options {
+  orientation: Orientation;
+}
